refactor(eks): split tools/__init__.py into one module per tool
