+++ tests/tools/test_eks_events_tool.py
@@ -51,7 +51,8 @@ def test_run_happy_path() -> None:
         ]
     )
     with patch(
-        "integrations.eks.tools.build_k8s_clients", return_value=(mock_core_v1, MagicMock())
+        "integrations.eks.tools.eks_events_tool.build_k8s_clients",
+        return_value=(mock_core_v1, MagicMock()),
     ):
         result = get_eks_events(
             cluster_name="c1", namespace="default", role_arn="arn:aws:iam::123:role/r"
@@ -66,7 +67,8 @@ def test_run_all_namespaces() -> None:
     mock_core_v1 = MagicMock()
     mock_core_v1.list_event_for_all_namespaces.return_value = MagicMock(items=[])
     with patch(
-        "integrations.eks.tools.build_k8s_clients", return_value=(mock_core_v1, MagicMock())
+        "integrations.eks.tools.eks_events_tool.build_k8s_clients",
+        return_value=(mock_core_v1, MagicMock()),
     ):
         result = get_eks_events(
             cluster_name="c1", namespace="all", role_arn="arn:aws:iam::123:role/r"
@@ -76,7 +78,10 @@ def test_run_all_namespaces() -> None:
 
 
 def test_run_handles_exception() -> None:
-    with patch("integrations.eks.tools.build_k8s_clients", side_effect=Exception("api error")):
+    with patch(
+        "integrations.eks.tools.eks_events_tool.build_k8s_clients",
+        side_effect=Exception("api error"),
+    ):
         result = get_eks_events(
             cluster_name="c1", namespace="default", role_arn="arn:aws:iam::123:role/r"
         )
