+++ tests/tools/test_eks_list_deployments_tool.py
@@ -46,7 +46,7 @@ def test_run_happy_path() -> None:
         items=[_make_deployment("dep-1"), _make_deployment("dep-2")]
     )
     with patch(
-        "integrations.eks.tools.build_k8s_clients",
+        "integrations.eks.tools.eks_list_deployments_tool.build_k8s_clients",
         return_value=(MagicMock(), mock_apps_v1),
     ):
         result = list_eks_deployments(
@@ -63,7 +63,7 @@ def test_run_detects_degraded() -> None:
         items=[_make_deployment("dep-1", desired=3, ready=2, unavailable=1)]
     )
     with patch(
-        "integrations.eks.tools.build_k8s_clients",
+        "integrations.eks.tools.eks_list_deployments_tool.build_k8s_clients",
         return_value=(MagicMock(), mock_apps_v1),
     ):
         result = list_eks_deployments(
@@ -76,7 +76,7 @@ def test_run_all_namespaces() -> None:
     mock_apps_v1 = MagicMock()
     mock_apps_v1.list_deployment_for_all_namespaces.return_value = MagicMock(items=[])
     with patch(
-        "integrations.eks.tools.build_k8s_clients",
+        "integrations.eks.tools.eks_list_deployments_tool.build_k8s_clients",
         return_value=(MagicMock(), mock_apps_v1),
     ):
         result = list_eks_deployments(
@@ -87,7 +87,10 @@ def test_run_all_namespaces() -> None:
 
 
 def test_run_handles_exception() -> None:
-    with patch("integrations.eks.tools.build_k8s_clients", side_effect=Exception("forbidden")):
+    with patch(
+        "integrations.eks.tools.eks_list_deployments_tool.build_k8s_clients",
+        side_effect=Exception("forbidden"),
+    ):
         result = list_eks_deployments(
             cluster_name="c1", namespace="default", role_arn="arn:aws:iam::123:role/r"
         )
