+++ tests/tools/test_eks_list_namespaces_tool.py
@@ -41,7 +41,7 @@ def test_run_happy_path() -> None:
         items=[_make_ns("default"), _make_ns("kube-system")]
     )
     with patch(
-        "integrations.eks.tools.build_k8s_clients",
+        "integrations.eks.tools.eks_list_namespaces_tool.build_k8s_clients",
         return_value=(mock_core_v1, MagicMock()),
     ):
         result = list_eks_namespaces(cluster_name="c1", role_arn="arn:aws:iam::123:role/r")
@@ -51,6 +51,9 @@ def test_run_happy_path() -> None:
 
 
 def test_run_handles_exception() -> None:
-    with patch("integrations.eks.tools.build_k8s_clients", side_effect=Exception("api error")):
+    with patch(
+        "integrations.eks.tools.eks_list_namespaces_tool.build_k8s_clients",
+        side_effect=Exception("api error"),
+    ):
         result = list_eks_namespaces(cluster_name="c1", role_arn="arn:aws:iam::123:role/r")
     assert result["available"] is False
