+++ tests/tools/test_eks_list_pods_tool.py
@@ -54,7 +54,8 @@ def test_run_happy_path() -> None:
         items=[_make_pod("pod-1"), _make_pod("pod-2")]
     )
     with patch(
-        "integrations.eks.tools.build_k8s_clients", return_value=(mock_core_v1, MagicMock())
+        "integrations.eks.tools.eks_list_pods_tool.build_k8s_clients",
+        return_value=(mock_core_v1, MagicMock()),
     ):
         result = list_eks_pods(
             cluster_name="c1", namespace="default", role_arn="arn:aws:iam::123:role/r"
@@ -70,7 +71,8 @@ def test_run_detects_crashing_pods() -> None:
         items=[_make_pod("pod-1", restart_count=10)]
     )
     with patch(
-        "integrations.eks.tools.build_k8s_clients", return_value=(mock_core_v1, MagicMock())
+        "integrations.eks.tools.eks_list_pods_tool.build_k8s_clients",
+        return_value=(mock_core_v1, MagicMock()),
     ):
         result = list_eks_pods(
             cluster_name="c1", namespace="default", role_arn="arn:aws:iam::123:role/r"
@@ -82,7 +84,8 @@ def test_run_all_namespaces() -> None:
     mock_core_v1 = MagicMock()
     mock_core_v1.list_pod_for_all_namespaces.return_value = MagicMock(items=[])
     with patch(
-        "integrations.eks.tools.build_k8s_clients", return_value=(mock_core_v1, MagicMock())
+        "integrations.eks.tools.eks_list_pods_tool.build_k8s_clients",
+        return_value=(mock_core_v1, MagicMock()),
     ):
         result = list_eks_pods(
             cluster_name="c1", namespace="all", role_arn="arn:aws:iam::123:role/r"
@@ -92,7 +95,10 @@ def test_run_all_namespaces() -> None:
 
 
 def test_run_handles_exception() -> None:
-    with patch("integrations.eks.tools.build_k8s_clients", side_effect=Exception("auth error")):
+    with patch(
+        "integrations.eks.tools.eks_list_pods_tool.build_k8s_clients",
+        side_effect=Exception("auth error"),
+    ):
         result = list_eks_pods(
             cluster_name="c1", namespace="default", role_arn="arn:aws:iam::123:role/r"
         )
