+++ tests/tools/test_eks_nodegroup_health_tool.py
@@ -42,7 +42,9 @@ def test_run_happy_path() -> None:
         "labels": {},
         "taints": [],
     }
-    with patch("integrations.eks.tools.EKSClient", return_value=mock_client):
+    with patch(
+        "integrations.eks.tools.eks_nodegroup_health_tool.EKSClient", return_value=mock_client
+    ):
         result = get_eks_nodegroup_health(cluster_name="c1", role_arn="arn:aws:iam::123:role/r")
     assert result["available"] is True
     assert len(result["nodegroups"]) == 1
@@ -61,7 +63,9 @@ def test_run_specific_nodegroup() -> None:
         "labels": {},
         "taints": [],
     }
-    with patch("integrations.eks.tools.EKSClient", return_value=mock_client):
+    with patch(
+        "integrations.eks.tools.eks_nodegroup_health_tool.EKSClient", return_value=mock_client
+    ):
         result = get_eks_nodegroup_health(
             cluster_name="c1", role_arn="arn:aws:iam::123:role/r", nodegroup_name="ng-specific"
         )
@@ -73,6 +77,8 @@ def test_run_handles_client_error() -> None:
     mock_client = MagicMock()
     error = ClientError({"Error": {"Code": "AccessDenied", "Message": "Denied"}}, "ListNodegroups")
     mock_client.list_nodegroups.side_effect = error
-    with patch("integrations.eks.tools.EKSClient", return_value=mock_client):
+    with patch(
+        "integrations.eks.tools.eks_nodegroup_health_tool.EKSClient", return_value=mock_client
+    ):
         result = get_eks_nodegroup_health(cluster_name="c1", role_arn="arn:aws:iam::123:role/r")
     assert result["available"] is False
