+++ tests/tools/test_telemetry.py
@@ -307,12 +307,12 @@ def _github_star_history_case() -> ToolFailureCase:
 
 def _eks_list_clusters_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_list_clusters_tool as mod
 
         mp.setattr(mod, "EKSClient", MagicMock(side_effect=RuntimeError("eks")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_list_clusters_tool as mod
 
         return mod.list_eks_clusters(role_arn="arn:aws:iam::123:role/x")
 
@@ -321,12 +321,12 @@ def _eks_list_clusters_case() -> ToolFailureCase:
 
 def _eks_describe_cluster_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_describe_cluster_tool as mod
 
         mp.setattr(mod, "EKSClient", MagicMock(side_effect=RuntimeError("eks")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_describe_cluster_tool as mod
 
         return mod.describe_eks_cluster(cluster_name="c", role_arn="arn:aws:iam::123:role/x")
 
@@ -335,12 +335,12 @@ def _eks_describe_cluster_case() -> ToolFailureCase:
 
 def _eks_nodegroup_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_nodegroup_health_tool as mod
 
         mp.setattr(mod, "EKSClient", MagicMock(side_effect=RuntimeError("eks")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_nodegroup_health_tool as mod
 
         return mod.get_eks_nodegroup_health(
             cluster_name="c",
@@ -352,12 +352,12 @@ def _eks_nodegroup_case() -> ToolFailureCase:
 
 def _eks_addon_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_describe_addon_tool as mod
 
         mp.setattr(mod, "EKSClient", MagicMock(side_effect=RuntimeError("eks")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_describe_addon_tool as mod
 
         return mod.describe_eks_addon(
             cluster_name="c",
@@ -370,12 +370,12 @@ def _eks_addon_case() -> ToolFailureCase:
 
 def _eks_events_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_events_tool as mod
 
         mp.setattr(mod, "build_k8s_clients", MagicMock(side_effect=RuntimeError("k8s")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_events_tool as mod
 
         return mod.get_eks_events(
             cluster_name="c",
@@ -388,12 +388,12 @@ def _eks_events_case() -> ToolFailureCase:
 
 def _eks_node_health_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_node_health_tool as mod
 
         mp.setattr(mod, "build_k8s_clients", MagicMock(side_effect=RuntimeError("k8s")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_node_health_tool as mod
 
         return mod.get_eks_node_health(
             cluster_name="c",
@@ -405,12 +405,12 @@ def _eks_node_health_case() -> ToolFailureCase:
 
 def _eks_list_namespaces_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_list_namespaces_tool as mod
 
         mp.setattr(mod, "build_k8s_clients", MagicMock(side_effect=RuntimeError("k8s")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_list_namespaces_tool as mod
 
         return mod.list_eks_namespaces(
             cluster_name="c",
@@ -422,12 +422,12 @@ def _eks_list_namespaces_case() -> ToolFailureCase:
 
 def _eks_list_deployments_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_list_deployments_tool as mod
 
         mp.setattr(mod, "build_k8s_clients", MagicMock(side_effect=RuntimeError("k8s")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_list_deployments_tool as mod
 
         return mod.list_eks_deployments(
             cluster_name="c",
@@ -440,12 +440,12 @@ def _eks_list_deployments_case() -> ToolFailureCase:
 
 def _eks_list_pods_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_list_pods_tool as mod
 
         mp.setattr(mod, "build_k8s_clients", MagicMock(side_effect=RuntimeError("k8s")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_list_pods_tool as mod
 
         return mod.list_eks_pods(
             cluster_name="c",
@@ -458,12 +458,12 @@ def _eks_list_pods_case() -> ToolFailureCase:
 
 def _eks_pod_logs_case() -> ToolFailureCase:
     def patch(mp: pytest.MonkeyPatch) -> None:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_pod_logs_tool as mod
 
         mp.setattr(mod, "build_k8s_clients", MagicMock(side_effect=RuntimeError("k8s")))
 
     def invoke() -> dict[str, Any]:
-        import integrations.eks.tools as mod
+        from integrations.eks.tools import eks_pod_logs_tool as mod
 
         return mod.get_eks_pod_logs(
             cluster_name="c",
@@ -867,7 +867,7 @@ def test_eks_client_error_path_uses_warning_severity(
     """
     from botocore.exceptions import ClientError
 
-    import integrations.eks.tools as mod
+    from integrations.eks.tools import eks_list_clusters_tool as mod
 
     client_error = ClientError(
         error_response={
@@ -917,7 +917,7 @@ def test_eks_nodegroup_health_tags_failing_nodegroup_during_iteration(
     failure on the second nodegroup should reach Sentry tagged with
     ``ng-broken``, not ``None`` or the first nodegroup.
     """
-    import integrations.eks.tools as mod
+    from integrations.eks.tools import eks_nodegroup_health_tool as mod
 
     def _describe(_cluster: str, ng: str) -> dict[str, Any]:
         if ng == "ng-broken":
